Player: optimize composite_player_vulnerability()
Remove a useless (and costly) school mask -> school type -> school mask
roundtrip when checking raid debuffs.

nv. Main cost comes from `dbc::school_type()` on the effects school
mask. As it's a bitmask it can't be optimized to a (jump)table and is
also practically unpredictable in the general case.