Micro-optimize sample data w/ min max
First, we don't need the `found` boolean. The observation is that if we
have at least one sample it necessarily has to be either lower than our
default min or higher than our default max. So we can simply key it off
the sample count.

Second, convert the setters into ternaries. On x64 this lets the
compiler generate `minsd`/maxsd` for the whole operation instead of
compare + jump.